Took the form out of App.jsx and made it its own component
